Fix limits on the fill function
Fill was only checking for the limit on line 150. This means it just dumped a response later regardless of the limit not being reached yet. I figure a quick recursive call to cause it to continue to fill until the limit is reached would get this working. It tests ok on my end but I only use the respond function.